Allow redis or redis cache adapters work with connection pool
