+++ lib/flipper/adapters/redis_shared/methods.rb
@@ -1,0 +1,9 @@
+module Flipper
+  module Adapters
+    module RedisShared
+      def with_connection(&block)
+        @client.respond_to?(:with) ? @client.with(&block) : yield(@client)
+      end
+    end
+  end
+end
